+++ src/App/Documentation/components/Loaders/index.js
@@ -1,7 +1,6 @@
 import React, { Component } from "react";
-import PrismCode from "react-prism";
 
-import { ComponentPreview, PxScript, Attribute, DocToc } from "#";
+import { ComponentPreview, Attribute, DocToc, Property } from "#";
 import LoaderComponent from "@/Loader";
 import { loader } from "$/px-script";
 
@@ -18,7 +17,7 @@ const BasicUsage = () => (
 const Sizes = () => (
     <>
         <h2 id="sizes">Sizes</h2>
-        <p>In addition to <code className="token property">.loader-default</code> you can use <code className="token property">.loader-small</code> and <code className="token property">.loader-large</code> to decrease or increase the size of the loader.</p>
+        <p>In addition to <Property value=".loader-default" /> you can use <Property value=".loader-small" /> and <Property value=".loader-large" /> to decrease or increase the size of the loader.</p>
         <h3>Small</h3>
         <ComponentPreview language="html" showCasePanel codeFigure>
             <LoaderComponent size="small" visible />
@@ -33,7 +32,7 @@ const Sizes = () => (
 const MutedLoader = () => (
     <>
         <h2 id="muted-loader">Muted loader</h2>
-        <p>Use class <code className="token property">.loader-muted</code> to mute the loader...</p>
+        <p>Use class <Property value=".loader-muted" /> to mute the loader...</p>
         <ComponentPreview language="html" showCasePanel codeFigure>
             <LoaderComponent id="test" size="default" visible muted />
         </ComponentPreview>
@@ -55,68 +54,6 @@ const StaticHtml = () => (
     </>
 );
 
-const UsageWithJavascript = () => (
-    <>
-        <h2 id="usage-with-javascript">Usage with javascript</h2>
-        <p>To interact with the loader-component with javascript...</p>
-        <h3>Show loaders</h3>
-        <ComponentPreview language="javascript" codeFigure>
-            {"px.loader.show();"}{"\n"}
-            {"// OR"}{"\n"}
-            {"px.loader.show(id);"}
-        </ComponentPreview>
-        <h3>Hide loaders</h3>
-        <ComponentPreview language="javascript" codeFigure>
-            {"px.loader.hide();"}{"\n"}
-            {"// OR"}{"\n"}
-            {"px.loader.hide(id);"}
-        </ComponentPreview>
-        <h3>Methods</h3>
-        <table className="table table-striped">
-            <thead>
-                <tr>
-                    <th>Method</th>
-                    <th>Description</th>
-                </tr>
-            </thead>
-            <tbody>
-                <tr>
-                    <td>
-                        <PxScript component="loader" func="show" />
-                    </td>
-                    <td>
-                        Gives all rendered loaders the class <code className="token property">.in</code>, effectively giving them the css property <PrismCode className="language-css">{"display: inline-block;"}</PrismCode>.
-                    </td>
-                </tr>
-                <tr>
-                    <td>
-                        <PxScript component="loader" func="show" params={["id"]} />
-                    </td>
-                    <td>
-                        Gives the specified loader the class <code className="token property">.in</code>, giving it the css property <PrismCode className="language-css">{"display: inline-block;"}</PrismCode>. If no loader with the specified id is rendered to the DOM, the method does nothing.
-                    </td>
-                </tr>
-                <tr>
-                    <td>
-                        <PxScript component="loader" func="hide" />
-                    </td>
-                    <td>
-                        Removes the class <code className="token property">.in</code> from all rendered loaders, effectively giving them the css property <PrismCode className="language-css">{"display: none;"}</PrismCode>
-                    </td>
-                </tr>
-                <tr>
-                    <td>
-                        <PxScript component="loader" func="hide" params={["id"]} />
-                    </td>
-                    <td>
-                        Removes the class <code className="token property">.in</code> from the loader specified by id, giving it the css property <PrismCode className="language-css">{"display: none;"}</PrismCode>. If no alert with the specified id is rendered to the DOM, the method does nothing.
-                    </td>
-                </tr>
-            </tbody>
-        </table>
-    </>
-);
-
 const LoadersText = () => (
     <div className="col-md-12 col-lg-10 doc-body">
         <p className="lead">Use loaders...</p>
@@ -124,7 +61,6 @@ const LoadersText = () => (
         <Sizes />
         <MutedLoader />
         <StaticHtml />
-        {/* <UsageWithJavascript /> */}
     </div>
 );
 
@@ -148,4 +84,4 @@ class Loaders extends Component {
 export default Loaders;
 
 /* For testing */
-export { BasicUsage, Sizes, MutedLoader, StaticHtml, UsageWithJavascript, LoadersText };
+export { BasicUsage, Sizes, MutedLoader, StaticHtml, LoadersText };