+++ src/App/Documentation/components/Tabs/index.js
@@ -2,7 +2,7 @@ import React, { Component } from "react";
 import { Link } from "react-router-dom";
 import PrismCode from "react-prism";
 
-import { ComponentPreview, DocToc, ExperimentalComponentAlert } from "#";
+import { ComponentPreview, DocToc, ExperimentalComponentAlert, Property } from "#";
 import { tabs } from "$/px-script";
 import TabsComponent from "@/Tabs";
 
@@ -11,7 +11,7 @@ const tabItems = ["Kort", "Rabatter", "Transaksjoner", "Faktura", "Instillinger"
 const BasicUsage = () => (
     <>
         <h2 id="basic-usage">Basic usage</h2>
-        <p>Add class <code className="token property">.tabs</code> with a <code className="token property">.tabs-horizontal-desired_size</code> to a div containing an arrow icon along with <PrismCode className="language-html">{"<ul>"}</PrismCode> and nest <PrismCode className="language-html">{"<li>"}</PrismCode> as needed. The <code className="token property">.tabs-horizontal-desired_size</code> class determines when your tabs will switch from horizontal to collapsed. The available sizes can be found in our <Link to="/docs/core/breakpoints">breakpoints documentation</Link>.</p>
+        <p>Add class <Property value=".tabs" /> with a <Property value=".tabs-horizontal-desired_size" /> to a div containing an arrow icon along with <PrismCode className="language-html">{"<ul>"}</PrismCode> and nest <PrismCode className="language-html">{"<li>"}</PrismCode> as needed. The <Property value=".tabs-horizontal-desired_size" /> class determines when your tabs will switch from horizontal to collapsed. The available sizes can be found in our <Link to="/docs/core/breakpoints">breakpoints documentation</Link>.</p>
         <ComponentPreview language="html" showCasePanel codeFigure>
             <TabsComponent mode="xl" items={tabItems}/>
         </ComponentPreview>
@@ -21,7 +21,7 @@ const BasicUsage = () => (
 const Collapsed = () => (
     <>
         <h2 id="collapsed">Collapsed</h2>
-        <p>If no <code className="token property">.tabs-horizontal-desired_size</code> is provided then the tabs will always be in collapsed mode.</p>
+        <p>If no <Property value=".tabs-horizontal-desired_size" /> is provided then the tabs will always be in collapsed mode.</p>
         <ComponentPreview language="html" showCasePanel codeFigure>
             <TabsComponent items={tabItems}/>
         </ComponentPreview>