+++ src/App/Documentation/utilities/Colors/index.js
@@ -1,12 +1,12 @@
 import React from "react";
 import PrismCode from "react-prism";
 
-import { ComponentPreview, DocToc } from "#";
+import { ComponentPreview, DocToc, Property } from "#";
 
 const TextColors = () => (
     <>
         <h2 id="text-colors">Text colors</h2>
-        <p>To use a color simply prepend <code className="token property">color-</code> to the desired color and add it to the element class attribute.</p>
+        <p>To use a color simply prepend <Property value="color-{name}" />.</p>
         <ComponentPreview language="html" showCasePanel codeFigure>
             <p className="color-brand">This text is color-brand.</p>
             <p className="color-default">This text is color-default.</p>
@@ -15,8 +15,8 @@ const TextColors = () => (
             <p className="color-warning">This text is color-warning.</p>
             <p className="color-danger">This text is color-danger.</p>
         </ComponentPreview>
-        <h3>With <code className="token property">.text-muted</code></h3>
-        <p>To get a muted effect on the color you are using, append a <PrismCode className="language-html">{"<span>"}</PrismCode>-element with the class <code className="token property">.text-muted</code> to the desired element.</p>
+        <h3>With <Property value=".text-muted" /></h3>
+        <p>To get a muted effect on the color you are using, append a <PrismCode className="language-html">{"<span>"}</PrismCode>-element with the class <Property value=".text-muted" /> to the desired element.</p>
         <ComponentPreview language="html" showCasePanel codeFigure>
             <p className="color-brand">This text is color-brand, <span className="text-muted">with a muted ending.</span></p>
             <p className="color-default">This text is color-default, <span className="text-muted">with a muted ending.</span></p>
@@ -31,7 +31,7 @@ const TextColors = () => (
 const BackgroundColors = () => (
     <>
         <h2 id="background-colors">Background colors</h2>
-        <p>To alter the background-color of an element, add the class <code className="token property">background-</code> appended with the desired color to get the desired background-color.</p>
+        <p>To alter the background-color of an element, add the class <Property value="background-{name}" />.</p>
         <ComponentPreview language="html" showCasePanel codeFigure>
             <div className="container">
                 <div className="row">