+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Added documentation section for text utlities.
+- Added documentation section for flex utlities.
+- Added documentation section for sizing utlities.
 - Added functionality to programmatically open and close dialog component.
 - Added functionality to programmatically open and close sheet component.
 
@@ -12,3 +15,5 @@
 ### Fixed
 
 ### Removed
+
+- Removed deprecated progress tracker component.